+++ firmware/cvc/test/test_Swc_Scheduler.c
@@ -18,7 +18,7 @@
 
 typedef unsigned char   uint8;
 typedef unsigned short  uint16;
-typedef unsigned long   uint32;
+typedef unsigned int   uint32;
 typedef uint8           Std_ReturnType;
 
 #define E_OK        0u
