+++ firmware/rzc/test/test_Swc_RzcCom.c
@@ -23,10 +23,10 @@
 
 typedef unsigned char   uint8;
 typedef unsigned short  uint16;
-typedef unsigned long   uint32;
+typedef unsigned int   uint32;
 typedef signed char     sint8;
 typedef signed short    sint16;
-typedef signed long     sint32;
+typedef signed int     sint32;
 typedef uint8           Std_ReturnType;
 
 #define E_OK        ((Std_ReturnType)0x00U)
